Updated CandyAdminController.java with tokens

+++ src/main/java/group3/kino/candyAdministration/controller/CandyAdminController.java
@@ -3,6 +3,7 @@ package group3.kino.candyAdministration.controller;
 
 import group3.kino.candyAdministration.model.Candy;
 import group3.kino.candyAdministration.service.CandyAdminService;
+import group3.kino.util.TokenAuthenticator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,19 +23,26 @@ public class CandyAdminController {
     }
 
     @PostMapping("/addCandy")
-    public ResponseEntity<Candy> addCandy(@RequestBody Candy candy){
+    public ResponseEntity<Candy> addCandy(@RequestBody Candy candy, @RequestHeader("token") int key) {
+        if (!TokenAuthenticator.isTokenAuthenticated(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         candyAdminService.save(candy);
         return new ResponseEntity<>(candy, HttpStatus.OK);
     }
 
     @GetMapping("/getAllCandy")
-    public ResponseEntity<Set<Candy>> getAllCandy(){
+    public ResponseEntity<Set<Candy>> getAllCandy(@RequestHeader("token") int key){
+        if (!TokenAuthenticator.isTokenAuthenticated(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
         return new ResponseEntity<>(candyAdminService.findAll(), HttpStatus.OK);
     }
 
 
     @PostMapping("/deleteCandy")
-    public ResponseEntity<Candy> deleteCandy(Long id){
+    public ResponseEntity<Candy> deleteCandy(Long id, @RequestHeader("token") int key){
+        if (!TokenAuthenticator.isTokenAuthenticated(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         Optional<Candy> candy = candyAdminService.findById(id);
         if (candy.isPresent()){
             Candy candy_ = candy.get();
@@ -46,12 +54,19 @@ public class CandyAdminController {
     }
 
     @DeleteMapping("/deleteCandy/{candyId}")
-    public void deleteCandyByID(@PathVariable("candyId") Long candyId){
+    public ResponseEntity<HttpStatus> deleteCandyByID(@PathVariable("candyId") Long candyId, @RequestHeader("token") int key){
+        if (!TokenAuthenticator.isTokenAuthenticated(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         candyAdminService.deleteById(candyId);
+        return new ResponseEntity<>(null, HttpStatus.OK);
     }
 
     @PutMapping("/editCandy/{candyId}")
-    public ResponseEntity<Candy> editCandy(@RequestBody Candy newCandy, @PathVariable() Long candyId) {
+    public ResponseEntity<Candy> editCandy(@RequestBody Candy newCandy, @PathVariable() Long candyId, @RequestHeader("token") int key) {
+        if (!TokenAuthenticator.isTokenAuthenticated(key))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         System.out.println("Candy: "+newCandy.getCandyName() + " \tId: " + candyId);
         Optional<Candy> oldCandy = candyAdminService.findById(candyId);
         if (oldCandy.isPresent()) {
